fixed all warnings, removed unused files

+++ lib/Graphics/src/Animation/Animation.cpp
@@ -17,7 +17,7 @@ namespace helperTools
                 return i; // index found, currently between i and i+1
             }
         }
-        return allKeyFrames.size() - 2; // between the second last and last already, return the second last
+        return static_cast<int>(allKeyFrames.size()) - 2; // between the second last and last already, return the second last
     }
 
     template <typename keyFrame>
@@ -73,9 +73,9 @@ namespace skeletonHelper
         {
             if (theSkeleton.bones[i].parentIndex == -1) 
             {
-                oldToNew[i] = orderedBones.size();  // remember: old index i to new index (current size)
+                oldToNew[i] = static_cast<int>(orderedBones.size());  // remember: old index i to new index (current size)
                 oneSkeletonBone bone = theSkeleton.bones[i];
-                bone.id = orderedBones.size();  // update bone ID to match new position
+                bone.id = static_cast<int>(orderedBones.size());  // update bone ID to match new position
                 // parentIndex stays -1 (it's root)
                 orderedBones.push_back(bone);
             }
@@ -96,9 +96,9 @@ namespace skeletonHelper
                 int oldBonesParentIndex = theSkeleton.bones[i].parentIndex;
                 if (oldBonesParentIndex == -1 || oldToNew[oldBonesParentIndex] != -1) // is this bone parent already in ordered list
                 {
-                    oldToNew[i] = orderedBones.size();
+                    oldToNew[i] = static_cast<int>(orderedBones.size());
                     oneSkeletonBone bone = theSkeleton.bones[i];
-                    bone.id = orderedBones.size();  // update bone ID to match new position
+                    bone.id = static_cast<int>(orderedBones.size());  // update bone ID to match new position
                     if (oldBonesParentIndex != -1)
                     {
                         bone.parentIndex = oldToNew[oldBonesParentIndex];  // Parent's NEW position
@@ -128,7 +128,7 @@ namespace skeletonHelper
         for (size_t i = 0; i < theSkeleton.bones.size(); i++)
         {
             std::set<int> visited;  // track where we've been
-            int current = i;        // start from this bone
+            int current = static_cast<int>(i);        // start from this bone
 
             // follow the parent chain
             while (current != -1)  // -1 means we reached the root
